Reuse docplex variables with matching name if found
This commit relies on the fact docplex variables are named according to
the formulae they represent.  If a variable already exists with the same
name a Formula object would give it, then it pertains to another such
object with exactly matching structure.  The truth values of these
formulae are necessarily the same, hence we might as well reuse the
variable in lieu of establishing separate constraints for either
structure.

The only downside is that subformulae (other than atoms) of the
redundant formula can remain unconfigured, thereby offering no direct
means of inspecting their corresponding optimized variables.