Substitute repr for str throughout op and theory
The Formula stringification code has always aimed to yield results that
could be pasted directly into the interpreter (assuming the existence of
propositions with appropriately matching names).  This behavior is a
better fit for repr than str, the latter of which instead serves to
reproduce the object in a more human-readable format.