Drop Not's implicit 0 operand after initialization
Before this commit, iterating over Not's operands was confusing and
error-prone because, in accordance with its implementation in terms of
Implies, Not kept a second, implicit 0 operand in addition to its more
conventional argument.  As a result, generic code to, e.g., clone an
operator could erroneously initialize Not with logic=0.

Even with this change, Not's super class continues to function because
Implies keeps its own copy of the implicit 0 in its rhs member variable.