Name constraints to avoid repetition on redundant Formulae
This commit adds names to constraints, which may then be checked to
prevent the addition of redundant constraints even when configuring
Formulae with exactly matching structure to other Formulae already
included in the docplex model.  As a result, configuration may safely
pass through such repeated Formulae and thus set Formula.val for all
subformulae in the tested Theory and/or query.