Define __eq__ for all subclasses of Formula
Formulae are equal if they have matching structure.  The motivation for
this change is the fact such Formulae now share docplex variables and
constraints even if included redundantly as subformulae throughout the
model, thereby rendering them functionally equivalent.

This commit also reinstates the default __hash__ wherever __eq__ is
defined so as to permit the use of Formulae in, e.g., dictionary keys.