Filter out double negation and contrapositives
This is done by changing the way Not is introduced into Formulae.
Specifically, rather than storing candidate subformulae with top-level
Not, Not is added when composing subformulae via Implies; as a result,
it can never occur in series.  Further, we eliminate contrapositives by
negating at most one of lhs and rhs, specifically the smaller or, if
they are the same size, earlier.  This works because lhs and rhs will
eventually swap, thereby testing the contrapositives of the elided
negations.