Fix __eq__ for cyclical Formulae
The reuse of docplex variables and constraints already relies on repr to
produce sufficiently unique strings to stand in for equality testing, so
__eq__ might as well do the same and reap the benefit of repr's support
for cyclical Formulae.

This commit also eliminates hashing for Formulae, which is at best
dubious for a mutable type and quite possibly bug-inducing for objects
comparing equal but nonetheless having different (memory address based)
hashes.  The upshot is that Formulae may no longer serve as keys in
dictionaries and the like, though of course their reprs still can.